Convert the -S options to have a mandatory options.
The -S option is the server option.  You have to specify an address to listen
on.  If you want to listen on any interface, you can use:

	0.0.0.0:port
	[::]:port

Given the address, we can figure out the family and create the appropriate
socket.

I have to move some routines around to avoid making prototypes.  I don't know
if this is a Good Thing, but it saves some typing.