Bug fix---compiler warnings.
On a 32-bit system, a [u]int64_t quantity can't be printed as a "long", but
as a "long long".  Also include a cast so we ensure we get the proper values
on a 32-bit stack.

Also fixed the definition of TWD_DEFAULT_DURATION to do all the calcuations
using long long.