Added outline of test_self.c
I strongly suspect many OSes and older versions will silently
twiddle the dscp and tos settings. So testing the self stuff
adds an innate host OS test and regression test. I'll argue
this is the first test we need to complete...

(but first up is the essential client/server udp plumbing code,
 then the protocol (but we can fudge that))

There is no point in writing the rest of this unless the
core tests can be made to work on a given OS.