Added "last seen" idea to RFC.
I wanted to be able to do continuous measurements, and also
cope with packet loss. So the return stream (by default
clocked at the same rate as the sending stream) can have up
to packet_size number of last seen acknowlegements.

On a good day, with a tail wind this means that summary data
will only need to encapsulate a final RTT's worth of
information.

The idea breaks down somewhat if we are going with
asymmetric tests, where you might get 20mbit down
and 4mbit. Even then, at a 6x1 ratio, you are
only sending 48 bytes extra in the last seen stanza,
and since we plan to pad out the packet to a minimal
value (probably over 120 bytes) anyway, at least this
ratio will be able to do full acknowlegements as they
happen.

haven't worked out much more than that