Basic pthreads example
ginned from the man page with a couple (but only a couple) on how
this will actually drive a test.

I'd like to have a C++ style static initializer per test_whatever.c.
Linux does this somehow (with MODULE_INIT) in C. So you'd have
callbacks defined in a struct and exported to the main driver
proggie.

Spent some time looking over how to set actual realtime privs but
haven't grokked it fully.