Better handling for the -S option.
Okay, so I switched the -S option back to being flag.  If it's set, and
there are no addresses listed on the command line, attempt to call
getsockname() on STDIN_FILENO, on the assumption that [x]inetd is starting
us.  If it succeeds, we can then check the sa_family of the returned address
to get the socket family.  Therefore, the --ipv4 and --ipv6 options are
not really needed and can be deleted.