ms resolution for several different clocks now easily achieved
(hah. Took all morning)

Still have to read the value to be *absolutely* sure under load,
but this is the basic outline of what needs to be done. Have to
watch for the clock wandering too... etc. But 2.5ms was what
I was aiming for and seem to have got below a us.