test_simple: mass of additional stuff flying in loose formation
it is unclear if setsockopt of tos/ttl/etc is needed on sending,
or it can be done properly with a fd or if a raw socket with
ip addrs and ports is needed.

it seems clear that setsockopt is needed for tcp transfers.

but the future is foggy. No wonder nobody does socket programming
at this level.